Splitting hidden attribute methods and add a couple of helper functions.
