+++ src/WmiScripting/Support/ApiObjects/AbstractWbemObject.php
@@ -3,6 +3,7 @@
 namespace PhpWinTools\WmiScripting\Support\ApiObjects;
 
 use PhpWinTools\Support\COM\ComVariantWrapper;
+use PhpWinTools\WmiScripting\Concerns\HasHiddenAttributes;
 use PhpWinTools\WmiScripting\Contracts\Jsonable;
 use PhpWinTools\WmiScripting\Contracts\Arrayable;
 use PhpWinTools\WmiScripting\Concerns\HasArrayableAttributes;
@@ -12,7 +13,8 @@ use PhpWinTools\WmiScripting\Concerns\HasArrayableAttributes;
  */
 class AbstractWbemObject implements Arrayable, Jsonable
 {
-    use HasArrayableAttributes;
+    use HasArrayableAttributes,
+        HasHiddenAttributes;
 
     protected $object;
